requirements.txt
python-telegram-bot==20.7
pymongo
sentence-transformers
faiss-cpu
numpy
